clean up comment endpoint
